+++ packages/avilatek_core/lib/avilatek_core.dart
@@ -1,4 +1,0 @@
-/// A Very Good Project created by Very Good CLI.
-library avilatek_core;
-
-export 'src/avilatek_core.dart';