+++ packages/avilatek_core/lib/common.dart
@@ -1,0 +1,5 @@
+/// Core utilities commonly used in Avila Tek projects.
+library core;
+
+export 'src/common/assertions.dart';
+export 'src/common/error_radar.dart';