+++ packages/avilatek_core/lib/http.dart
@@ -1,0 +1,7 @@
+/// HTTP client and utilities commonly used in Avila Tek projects.
+library http;
+
+export 'interfaces.dart' show DeviceStorage;
+export 'src/http/headers_injector.dart';
+export 'src/http/persistent_headers_injector.dart';
+export 'src/http/session_client.dart';