+++ packages/avilatek_core/lib/interfaces.dart
@@ -1,0 +1,5 @@
+/// Interfaces for most common use cases in Avila Tek projects, and
+/// for other πackages of avilatek_core.
+library interfaces;
+
+export 'src/interfaces/device_storage.dart';